\name{demo.NO}
\alias{demo.NO}
\alias{demo.LO}
\alias{demo.NO.LO}
\alias{demo.GU}
\alias{demo.RG}
\alias{demo.exGAUS }
\alias{demo.PE}
\alias{demo.PE.NO}
\alias{demo.TF}
\alias{demo.TF.NO}
\alias{demo.EGB2}
\alias{demo.GT}
\alias{demo.JSU}
\alias{demo.JSUo}
\alias{demo.SHASH}
\alias{demo.NET}
\alias{demo.SEP1}
\alias{demo.SEP2}
\alias{demo.SEP3}
\alias{demo.SEP4}
\alias{demo.ST1}
\alias{demo.ST2}
\alias{demo.ST3}
\alias{demo.ST4}
\alias{demo.ST5}
\alias{demo.EXP}
\alias{demo.GA}
\alias{demo.LOGNO}
\alias{demo.NO.LOGNO}
\alias{demo.WEI}
\alias{demo.WEI2}
\alias{demo.WEI3}
\alias{demo.BCCG}
\alias{demo.GG}
\alias{demo.GIG}
\alias{demo.ZAGA}
\alias{demo.ZAIG}
\alias{demo.BCT}
\alias{demo.BCPE}
\alias{demo.GB2}
\alias{demo.BE}
\alias{demo.BEo} 
\alias{demo.GB1}  
\alias{demo.BB}
\alias{demo.BEINF}
\alias{demo.BEINF0}
\alias{demo.BEINF1}
\alias{demo.BI}
\alias{demo.DEL}
\alias{demo.LG}
\alias{demo.NBI}
\alias{demo.NBII}
\alias{demo.PO}
\alias{demo.SICHEL}
\alias{demo.ZABI}
\alias{demo.ZAGA}
\alias{demo.ZALG}
\alias{demo.ZAP}
\alias{demo.ZIBI}
\alias{demo.ZIP}
\alias{demo.ZIP2}
\alias{demo.IG}
\alias{demo.WEI}
\alias{demo.PIG}
\alias{demo.ZABB}
\alias{demo.ZIBB}
\alias{demo.ZANBI}
\alias{demo.ZINBI}
\alias{demo.ZIPIG}
\alias{demo.PIG}
\alias{demo.NOtr}
\alias{demo.GAtr}


\title{Demos for different gamlss.family distributions}
\description{
The demo functions for showning the \code{gamlss.family} distributions. The functions use the package \code{Rpanel}. 
}
\usage{
demo.NO()
demo.LO()
demo.NO.LO()
demo.GU()
demo.RG()
demo.exGAUS()
demo.PE()
demo.PE.NO()
demo.TF()
demo.TF.NO()
demo.EGB2()
demo.GT()
demo.JSU()
demo.JSUo()
demo.NET()
demo.SHASH()
demo.SEP1()
demo.SEP2()
demo.SEP3() 
demo.SEP4()
demo.ST1()
demo.ST2()
demo.ST3()
demo.ST4()
demo.ST5()   
demo.EXP()
demo.GA()
demo.LOGNO()
demo.NO.LOGNO()
demo.IG()
demo.WEI()
demo.WEI2()
demo.WEI3()
demo.BCCG()
demo.GG()
demo.GIG()
demo.ZAGA()
demo.ZAIG()
demo.BCT() 
demo.BCPE()
demo.GB2()
demo.EGB2()
demo.BE()
demo.BEo() 
demo.GB1()  
demo.GT()
demo.BB()
demo.BEINF()
demo.BEINF0()
demo.BEINF1()
demo.BI()
demo.DEL()
demo.LG()
demo.NBI()
demo.NBII()
demo.PO()
demo.SICHEL()
demo.ZABI()
demo.ZAGA()
demo.ZALG()
demo.ZAP()
demo.ZIBI()
demo.ZIP()
demo.ZIP2()
demo.BCCG()
demo.GG()
demo.PIG()
demo.ZABB()
demo.ZIBB()
demo.ZANBI()
demo.ZINBI()
demo.ZIPIG()
demo.NOtr() 
demo.GAtr()
}

\value{An rpanel plot}
\references{
Bowman, Bowman, Gibson and Crawford (2008) rpanel, CRAN

Rigby, R. A. and  Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), 
\emph{Appl. Statist.}, \bold{54}, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R.
Accompanying documentation in the current GAMLSS  help files, (see also  \url{http://www.gamlss.org/}).  

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R.
\emph{Journal of Statistical Software}, Vol. \bold{23}, Issue 7, Dec 2007, \url{http://www.jstatsoft.org/v23/i07}.

}

\author{
Mikis Stasinopoulos \email{d.stasinopoulos@londonmet.ac.uk}, Bob Rigby \email{r.rigby@londonmet.ac.uk} 
with contribution from Larisa Kosidou. 
}



\examples{
demo.NO()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{distribution}
\keyword{regression}
