\name{demoPsplines}
\alias{demoPsplines}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Interface for demonstrating the P-splines and other smoothers}
\description{
The function \code{demoPsplines} is an \code{tcltk} interface  for P. Eilers and B. Marx demos for P-splines.
}
\usage{
demoPsplines()
}

\value{Create an \code{tcltk} menu}
\references{
R Development Core Team (2010) tcltk package, CRAN. 

Bowman, Bowman, Gibson and Crawford (2008) rpanel, CRAN

Rigby, R. A. and  Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), 
\emph{Appl. Statist.}, \bold{54}, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R.
Accompanying documentation in the current GAMLSS  help files, (see also  \url{http://www.gamlss.org/}).  

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R.
\emph{Journal of Statistical Software}, Vol. \bold{23}, Issue 7, Dec 2007, \url{http://www.jstatsoft.org/v23/i07}.
}
\author{
Konstantinos Pateras \email{kostas.pateras@gmail.com} 
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See also \code{\link{demoDist}}, ~~~
}
\examples{
demoPsplines()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{regression}

