\name{gamlss.demo}
\alias{gamlss.demo}
\title{The demo for gamlss distributions and smoothing}
\description{
It starts the gamlss demos. It is an \code{tcltk} interface for using the gamlss demos.
}
\usage{
gamlss.demo()
}


\value{
It creates a \code{tcltk} menu
}
\references{
R Development Core Team (2010) tcltk package, CRAN. 

Bowman, Bowman, Gibson and Crawford (2008) rpanel, CRAN

Rigby, R. A. and  Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), 
\emph{Appl. Statist.}, \bold{54}, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R.
Accompanying documentation in the current GAMLSS  help files, (see also  \url{http://www.gamlss.org/}).  

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R.
\emph{Journal of Statistical Software}, Vol. \bold{23}, Issue 7, Dec 2007, \url{http://www.jstatsoft.org/v23/i07}.

}
\author{
Konstantinos Pateras \email{kostas.pateras@gmail.com} 
}

\seealso{
See also \code{\link{demoDist}}, \code{\link{gamlss.demo}},
}
\examples{
gamlss.demo()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{regression}
